requirements.txt
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
timm==1.0.12
torch==2.2.2+cu118
torchvision==0.17.2+cu118
